* commit '03c2012328c9e289abd579c0bdf2725011978dba':
  Changed variables to float
  More minor formatting changes
  Minor formatting changes
  Small formatting changes
  Modified Readme
  Modify readme.md
  Modify readme.md
  Add images

+++ src/main/java/ca/uwo/csd/cs2212/team08/Goal.java
@@ -8,7 +8,7 @@ public class Goal implements Serializable {
 
 	private static final long serialVersionUID = -5343638548252582826L;
 	
-	private int target;
+	private float target;
     private boolean achieved;
     private String type;
 
@@ -18,7 +18,7 @@ public class Goal implements Serializable {
     }
 
     /*Getters and Setters */
-    public int getTarget() {
+    public float getTarget() {
         return target;
     }
 
@@ -53,12 +53,13 @@ public class Goal implements Serializable {
     	this.type= "floorsclimbed";
     	this.target = target;
     }
-    
+    /*
     public void setActiveMinutes(int target){
-    	this.type= "activeminutes";
-    	this.target = target;
+        this.type= "activeminutes";
+        this.target = target;
     }
-    
+    */
+
     public void setSedentaryMinutes(int target){
     	this.type= "sedentaryminutes";
     	this.target = target;
